double in dev

+++ src/main/java/igatorpackage/MyCompany.java
@@ -5,7 +5,7 @@ public class MyCompany {
 	private String name;
 	private String address;
 	private String company;
-	private int salary;
+	private double salary;
 	private int age;
 	private int experience;
 	private String surName;
@@ -29,7 +29,7 @@ public class MyCompany {
 	public void setCompany(String company) {
 		this.company = company;
 	}
-	public int getSalary() {
+	public double getSalary() {
 		return salary;
 	}
 	public void setSalary(int salary) {
